+++ src/main/java/Itstime/planear/member/dto/CreateMemberResponse.java
@@ -1,4 +1,4 @@
-package Itstime.planear.member.controller;
+package Itstime.planear.member.dto;
 
 public record CreateMemberResponse(String message, Long id) {
 }
